WFLY-13436 Combine AbstractDeploymentUnitProcessor and AnnotatedEJBComponentDescriptionDeploymentUnitProcessor, changing some protected methods to private.
