+++ ejb3/src/main/java/org/jboss/as/ejb3/deployment/processors/EJBComponentDescriptionFactory.java
@@ -21,7 +21,7 @@
  */
 package org.jboss.as.ejb3.deployment.processors;
 
-import static org.jboss.as.ejb3.deployment.processors.AbstractDeploymentUnitProcessor.getEjbJarDescription;
+import static org.jboss.as.ejb3.deployment.processors.AnnotatedEJBComponentDescriptionDeploymentUnitProcessor.getEjbJarDescription;
 
 import org.jboss.as.ee.component.Attachments;
 import org.jboss.as.ejb3.component.EJBComponentDescription;
